pyproject.toml
[tool.poetry]
name = "shop"
version = "0.1.0"
description = ""
authors = ["himokkk <himok.pk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
psycopg2 = "^2.9.3"
djangorestframework = "^3.14.0"
Pillow = "^9.2.0"
black = "^22.8.0"
isort = "^5.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
